remove duplicate (and confusing) wait_for method
we had both a `wait_for` and a `wait_for_element` method, one had it’s
own retry, while the other passed that responsibility to the ByClause.

I’ve decided the ByClause is responsible for the behavior here, and
`wait_for_element` seemed too verbose given the element definition are
the required arguments. This also officially makes `wait_for` just a
wrapper around passing the browser to a ByClause.